requirements.txt
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
tqdm>=4.66
Pillow>=10.0
opencv-python>=4.8
torch>=2.1
torchvision>=0.16
facenet-pytorch>=2.5
datasets>=2.18
huggingface_hub>=0.20
